+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KREDMNRTQ3AS3H4EHBH9E0NT"
+        "01KRG5HGST3QN1HV93A6SFJV0J"
     )
     r"""The unique identifier of the data source"""