+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KREDMNRPYA1FNNP5NCE26Z8R"
+        "01KRG5HGSRN4TMG7F7C2TQM2C2"
     )
     r"""The unique identifier of the data source"""