+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -154,7 +154,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-12T15:41:56.904Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-13T07:58:53.577Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")