+++ packages/orq-rc/RELEASES.md
@@ -7859,3 +7859,13 @@ Based on:
 - [python v4.9.0-rc.53] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.53] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.53 - packages/orq-rc
+
+## 2026-05-13 08:01:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.0] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.0] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.0 - packages/orq-rc